Fix PATH resolution issue
Dispatch of the sub command was failing using the bundle on MacOS, but
oddly not on Linux (use of vendored gems maybe). Caused cryptic errors:

    bundle exec bin/cidb scan
    Traceback (most recent call last):
            1: from /Users/markaddison/.gem/ruby/2.7.0/bin/cidb-scan:23:in `<main>'
    /Users/markaddison/.gem/ruby/2.7.0/bin/cidb-scan:23:in `load': /Users/markaddison/Lab/cidb/bin/cidb-scan:4: syntax error, unexpected unary-, expecting `do' or '{' or '(' (SyntaxError)
    set -e -o pipefail

Cuased when the gem bin stubs directory is in the PATH before the bin in
the repo. The generated stubs assume ruby (ignoring the shebang) and try
and run our bash as ruby, with the above error.

Gem executables may not be the best way to manage these bin files.
Seperate libexec probably needed.