Refactor JUnit classes
Rename JUnit::Sax to JUnit::Parser and add parse method to make it
clearer what is going on here, especially at call sites. They don't need
to know the Ox detail here, just how to write a handler and parse
a file.

Move the IsJUnitSax thing into the command for now, until (if we ever
need to) clean it up more or use else where.