Rework Sloppy#fail! into Logging#fatal!
Have the fail in the command line lib was getting a bit messy and felt
like exit was in the wrong place. We can get fatal errors in the libs as
well and don't want unintended exit calls.

Instead have a FatalError we throw after logging the error and have the
command runner catch that and exit gracefully. Keeps the code closer
together and avoids hidden exits, exceptions is how ruby does this.